Modi Mishaal

+++ restaurants/views.py
@@ -3,6 +3,9 @@ from .models import Restaurant, Item
 from .forms import RestaurantForm, ItemForm, SignupForm, SigninForm
 from django.contrib.auth import login, authenticate, logout
 
+def no_access(request):
+       return render(request, 'no-access.html')
+
 def signup(request):
     form = SignupForm()
     if request.method == 'POST':
@@ -59,6 +62,9 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+
+    if request.user.is_anonymous:
+        return redirect('signin')
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -75,6 +81,10 @@ def restaurant_create(request):
 def item_create(request, restaurant_id):
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+
+    if not (request.user.is_staff or request.user == restaurant.owner ):
+        return redirect('no-access')
+
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -89,7 +99,12 @@ def item_create(request, restaurant_id):
     return render(request, 'item_create.html', context)
 
 def restaurant_update(request, restaurant_id):
+
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+
+    if not (request.user.is_staff or request.user == restaurant_obj.owner ):
+        return redirect('no-access')
+
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
@@ -103,6 +118,9 @@ def restaurant_update(request, restaurant_id):
     return render(request, 'update.html', context)
 
 def restaurant_delete(request, restaurant_id):
+    
+    if not request.user.is_staff:
+        return redirect('no-access')
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
